migracije, dropdown club id, button na edit blade

+++ database/migrations/2019_10_07_175721_create_selection_table.php
@@ -19,8 +19,8 @@ class CreateSelectionTable extends Migration
             $table->string('gender');
             $table->string('category');
             $table->string('history');
-            $table->integer('staffTypeId');
-            $table->integer('clubId');
+            $table->bigInteger('staffTypeId');
+            $table->bigInteger('clubId');
             $table->timestamps();
 
             $table->foreign('staffTypeId')->references('id')->on('staff_types');
